# spec/dashboard/games_openapi.yaml
openapi: 3.0.0
info:
  description: "Qilin api "
  version: "1.0.0"
  title: "Qilin games dashboard API"
  contact:
    email: "roman.golenok@protocol.one"
...
servers:
  - url: https://qilinapi.tst.protocol.one
    description: Master test server
paths:
  /api/v1/genres:
    get:
      tags:
        - "common"
      summary: "Get list of genre"
      parameters:
        - name: title
          in: query
          description: "Filter genre by title"
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/tags:
    get:
      tags:
        - "common"
      summary: "Get list of tags"
      parameters:
        - name: title
          in: query
          description: "Filter tags by title"
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/vendor/:id/games:
    post:
      tags:
        - "game"
      summary: "Create game"
      operationId: "createGame"
      requestBody:
        description: "Internal name of game"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                internalName:
                  type: string
                  example: "MEGA_PUBG_CLONE"
                vendorId:
                  type: string
                  format: uuid
                  description: "Vendor who owns the game"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameInfo'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - "game"
        - "vendor"
      summary: "Get list of games in vendor"
      operationId: "listGames"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of vendor"
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - name: "internalName"
          in: query
          description: "Filter games by internalName"
          schema:
            type: "string"
          required: false
        - name: "genre"
          in: query
          description: "Filter games by genre"
          schema:
            type: "string"
          required: false
        - name: "price"
          in: query
          description: "Filter games by price"
          schema:
            type: "number"
            format: float
          example: 59.99
          required: false
        - name: "releaseDate"
          in: query
          description: "Filter games by release date"
          schema:
            type: "string"
            format: "date-time"
          example: "2017-07-21T17:32:28Z"
          required: false
        - name: "sort"
          in: query
          description: "Game name query filter"
          example: -genre,-releaseDate,+name
          schema:
            type: "array"
            items:
              type: "string"
              enum:
                - -genre
                - +genre
                - -releaseDate
                - +releaseDate
                - -price
                - +price
                - -internalName
                - +internalName
          required: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ShortGameInfo"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/games/{id}:
    get:
      tags:
        - "game"
      summary: "Get game info"
      description: ""
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          required: true
          description: "Id of game"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameInfo"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'


    delete:
      tags:
        - "game"
      summary: "Delete game"
      description: ""
      operationId: "deleteGame"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          required: true
          description: "Id of game to remove"
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - "game"
      summary: "Change game info"
      description: ""
      operationId: "changeGame"
      requestBody:
        description: "Game object that needs to be saved in store"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeGame"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
            format: uuid
          required: true
          description: "Id of game to change"
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/games/{id}/descriptions:
    get:
      tags:
        - game
      summary: "Gets additional descriptions of game"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: true
          example: "2019-01-11T17:32:28Z"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Description"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - game
      summary: "Change descriptions"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: string
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Description"
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
...
  /api/v1/descriptors:
    get:
      tags:
        - common
      summary: Gets list of descriptors
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Descriptor'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal Server Error
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 409
            message: "Resource with same unique id already exists"
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 422
            message: "Fields required: 'title, description'"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    Offset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: The number of items to skip before starting to collect the result set.
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
      description: The numbers of items to return.
  schemas:
...
    Descriptor:
      type: object
      properties:
        id:
          type: integer
        title:
          type: object
          additionalProperties:
            type: string
          example:
            en: Blood
            ru: Кровь
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Description:
      type: object
      properties:
        tagline:
          type: object
          additionalProperties:
            type: string
          example:
            en: "Short description"
            ru: "Короткое описание"
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en: "Full Description"
            ru: "полное описание"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        additionalDescription:
          type: string
        gameSite:
          type: string
          format: url
        socials:
          type: object
          items:
            type: string
          example:
            facebook: "https://facebook.com/yougame"
            twitter: "https://twitter.com/yougame"
...
    Tag:
      type: object
      properties:
        id:
          type: "string"
          readOnly: true
        title:
          type: object
          additionalProperties:
            type: string
          readOnly: true
          example:
            en: multiplayer
            ru: мультиплеер
    GameGenres:
      type: object
      properties:
        main:
          type: integer
        addition:
          type: array
          items:
            type: integer
    Genre:
      type: object
      properties:
        id:
          type: "string"
          readOnly: true
        title:
          type: object
          additionalProperties:
            type: string
          example:
            en: Action
            ru: Экшен
          readOnly: true
...
    ShortGameInfo:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        internalName:
          type: "string"
          readOnly: true
        icon:
          type: "string"
          readOnly: true
        genres:
          $ref: '#/components/schemas/GameGenres'
          readOnly: true
        releaseDate:
          type: "string"
          format: "date-time"
          readOnly: true
        prices:
          $ref: "#/components/schemas/Price"
    Price:
      type: object
      properties:
        price:
          type: number
          format: float
        currency:
          type: string
      example:
        price: 59.99
        currency: USD
    BaseGame:
      type: object
      properties:
        internalName:
          type: "string"
          readOnly: true
        # non-localized title of game for end user
        title:
          type: "string"
        # comma separated list of developers
        developers:
          type: "string"
        # comma separated list of publishers
        publishers:
          type: "string"
        releaseDate:
          type: "string"
          format: "date-time"
        displayRemainingTime:
          type: "boolean"
          default: false
        features:
          $ref: "#/components/schemas/Features"
        platforms:
          $ref: "#/components/schemas/SupportPlatform"
        requirements:
          type: object
          properties:
            macOs:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
            linux:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
            windows:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
        languages:
          $ref: "#/components/schemas/Languages"
        achievementOnProd:
          type: boolean
          default: false
          description: "Enables achievements on production"
    GameInfo:
      allOf:
        - $ref: '#/components/schemas/BaseGame'
        - type: object
          properties:
            genres:
              $ref: '#/components/schemas/GameGenres'
            tags:
              type: array
              items:
                type: integer
    ChangeGame:
      allOf:
        - $ref: '#/components/schemas/BaseGame'
        - type: object
          properties:
            genres:
              $ref: '#/components/schemas/GameGenres'
            tags:
              type: array
              items:
                type: integer
...
security:
  - bearerAuth: []